% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{calculate_priori}
\alias{calculate_priori}
\title{Generates vector of priori values one for each row in x_y_rec_checks,
This function is the default for creating priori values but,..
You can inject your own priori functions by passing a function into the priori_func parameter into the find_posterior function.
this function should be able to take all the parameters that this function takes}
\usage{
calculate_priori(
  x_y_rec_checks,
  n_x,
  n_y,
  total_comparisons = NULL,
  row_name_nm = "row_name",
  suffix = c("x", "y"),
  ...
)
}
\arguments{
\item{x_y_rec_checks}{Dataframe with records indicating row_numbers of each of the datasets.}

\item{n_x}{Integer. Number of records in first dataset}

\item{n_y}{Integer. Number of records in second dataset}

\item{total_comparisons}{Integer. Indicates total Comparisons that could be done between datasets#'}

\item{row_name_nm}{String. rowname to use for each dataset. Default 'row_name'}

\item{suffix}{String vector of length 2. Default c('x', 'y')}
}
\description{
Generates vector of priori values one for each row in x_y_rec_checks,
This function is the default for creating priori values but,..
You can inject your own priori functions by passing a function into the priori_func parameter into the find_posterior function.
this function should be able to take all the parameters that this function takes
}
\examples{
calculate_priori(
                 tibble(row_name.x = sample(1:1000, 100, replace=T),
                        row_name.y = sample(1:1000000, 100, replace=T)),
                 n_x = 100,
                 n_y = 1000000
)

}
