% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{clean_str_2}
\alias{clean_str_2}
\title{Cleans a string using after it has been tokenized as a like final step. This is the the default string cleaner used after tokenization
It can be overriden in tokenizer_basic, tokenize_col, tokenize_df, etc by passing a new function as post_token_clean_str.}
\usage{
clean_str_2(x, ...)
}
\arguments{
\item{x}{vector of strings}

\item{...}{ignored, used to ensure pass by keyword}
}
\description{
Cleans a string using after it has been tokenized as a like final step. This is the the default string cleaner used after tokenization
It can be overriden in tokenizer_basic, tokenize_col, tokenize_df, etc by passing a new function as post_token_clean_str.
}
\examples{
c('Z.Y. do things inc', 'z. y. DO things montrèal', 'at&t') |> clean_str_2(token_type = 'company_name')

}
