% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{find_posterior_all_evidence}
\alias{find_posterior_all_evidence}
\title{t_dat should have been run through find_posterior_positive_evidence_only() first to create the list of candidates}
\usage{
find_posterior_all_evidence(
  t_dat,
  min_posterior = MIN_POSTERIOR_ALL_EVIDENCE,
  token_join_by = TOKEN_TOKEN_TYPE_VEC,
  positive_evidence_only = NULL,
  return_all = TRUE,
  priori_func = calculate_priori,
  ...
)
}
\arguments{
\item{t_dat}{t_dat object}

\item{min_posterior}{Filter results below this threshold Default = 0.01}

\item{token_join_by}{What to join the all_tokens by. Default = TOKEN_TOKEN_TYPE_VEC,}

\item{positive_evidence_only, }{NULL or a dataframe with at least two columns indicating the row_names to be checked in each of the datasets}

\item{return_all}{If TRUE it returns the whole object if FALSE it just returns the dataframe. Default TRUE}

\item{priori_func}{A function that will calculate the priori. Default calculate_priori}

\item{...}{Not used}
}
\description{
t_dat should have been run through find_posterior_positive_evidence_only() first to create the list of candidates
}
\examples{


t_dat <-
  token_links(
     dat_x = readr::read_csv('https://tinyurl.com/2p8etjr6'),
     dat_y = readr::read_csv('https://tinyurl.com/2p8ap4ad'),
     args_x = list(col_nms = 'coname'),
     args_y = list(col_nms = 'companyName'),
     token_types = 'company_name',
     token_index = '',
     suffix = c('ceo', 'alb')
  )

results <-
  t_dat |>
    find_posterior_positive_evidence_only() |>
    find_posterior_all_evidence(return_all = FALSE)


}
