% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_posterior_chunked.R
\name{find_posterior_chunked}
\alias{find_posterior_chunked}
\title{Like "find_posterior" but it uses less memory at any one time by iterating over chunks of the x dataset.
When this is used the returned t_dat object will not have "positive_evidence_only" dataframe it will have the "all_evidence" dataframe}
\usage{
find_posterior_chunked(
  t_dat,
  ...,
  min_posterior_all_evidence = MIN_POSTERIOR_ALL_EVIDENCE,
  min_posterior_positive_evidence_only = MIN_POSTERIOR_POSITIVE_EVIDENCE_ONLY,
  n_chunks = N_CHUNKS,
  return_all = TRUE
)
}
\arguments{
\item{t_dat}{t_dat object}

\item{...}{passed to find_posterior_subset}

\item{min_posterior_all_evidence}{passed to find_posterior_subset. : Default MIN_POSTERIOR_ALL_EVIDENCE}

\item{min_posterior_positive_evidence_only}{passed to find_posterior_subset. : Default MIN_POSTERIOR_POSITIVE_EVIDENCE_ONLY}

\item{n_chunks}{How many chunks to break up the x dataset into : Default 100}

\item{return_all}{return modified "t_dat" object if TRUE otherwise return the all_evidence dataframe : Default TRUE}
}
\description{
Like "find_posterior" but it uses less memory at any one time by iterating over chunks of the x dataset.
When this is used the returned t_dat object will not have "positive_evidence_only" dataframe it will have the "all_evidence" dataframe
}
