% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{find_posterior_positive_evidence_only}
\alias{find_posterior_positive_evidence_only}
\title{Creates a subset of pairs to check in more detail.}
\usage{
find_posterior_positive_evidence_only(
  t_dat,
  min_posterior = MIN_POSTERIOR_POSITIVE_EVIDENCE_ONLY,
  token_join_by = TOKEN_TOKEN_TYPE_VEC,
  tokens_to_keep = NULL,
  return_all = TRUE,
  priori_func = calculate_priori,
  remove_identical_row_index = FALSE,
  x_rows_filter = NULL,
  y_rows_filter = NULL,
  ...
)
}
\arguments{
\item{t_dat}{t_dat object}

\item{min_posterior}{filter posterior results above this value. Default 0.2.}

\item{token_join_by}{vector column names that joins the tokens. Default TOKEN_TOKEN_TYPE_VEC}

\item{tokens_to_keep}{NULL or dataframe with a list of tokens, and m_prob and u_prob for each token, in the case where it is NULL we use tokenized$tokens_all filtered by min_token_u_prob}

\item{return_all}{if TRUE it returns the whole object if FALSE it just returns the dataframe. Default TRUE}

\item{priori_func}{A function that will calculate the priori. Default calculate_priori}

\item{...}{passed to keep_tokens() if tokens_to_keep is NULL}
}
\description{
Creates a subset of pairs to check in more detail.
}
\examples{
token_links(
 dat_x = readr::read_csv('https://tinyurl.com/2p8etjr6'),
 dat_y = readr::read_csv('https://tinyurl.com/2p8ap4ad' ),
 args_x = list(col_nms = 'coname'),
 args_y = list(col_nms = 'companyName'),
 token_types = 'company_name',
 token_index = '',
 suffix = c('ceo', 'alb')
) |> find_posterior_positive_evidence_only(return_all = FALSE)

}
