% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_posterior_chunked.R
\name{find_posterior_subset}
\alias{find_posterior_subset}
\title{Like "find_posterior" but this will always set "return_all = FALSE" and will require a subset of rownames for at least the X data set. but also maybe the y dataset}
\usage{
find_posterior_subset(
  t_dat,
  x_rows_filter,
  ...,
  y_rows_filter = NULL,
  min_posterior_all_evidence = NULL,
  min_posterior_positive_evidence_only = NULL
)
}
\arguments{
\item{t_dat}{t_dat object}

\item{x_rows_filter}{vector of row numbers to filter the x dataset by NO Default}

\item{...}{passed to both find_posterior_positive_evidence_only and find_posterior_all_evidence}

\item{y_rows_filter}{vector of row numbers to filter the y dataset Default NULL}

\item{min_posterior_all_evidence}{min_posterior_all_evidence passed to find_posterior_all_evidence as min_posterior. Default NULL}

\item{min_posterior_positive_evidence_only}{passed to find_posterior_positive_evidence_only as min_posterior. Default NULL}
}
\description{
Like "find_posterior" but this will always set "return_all = FALSE" and will require a subset of rownames for at least the X data set. but also maybe the y dataset
}
