% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{generate_all_tokens}
\alias{generate_all_tokens}
\title{Generates a dataframe with the total counts of each tokens across both datasets as well as the m and u probs}
\usage{
generate_all_tokens(
  ...,
  x_counts,
  y_counts,
  total_comparisons,
  token_count_join = TOKEN_TOKEN_TYPE_VEC,
  suffix = TOKEN_SUFFIX_DEFAULT,
  m_prob_func = calc_m_prob
)
}
\arguments{
\item{...}{not used}

\item{x_counts}{Counts of tokens from first dataset}

\item{y_counts}{Counts of tokens from second dataset}

\item{total_comparisons}{count of the number of comparisons that can happens normally is nrow(x_dat) * nrow(y_dat)}

\item{token_count_join}{String vector that joins the two token count dataframes. Default c('token','token_type')}

\item{suffix}{String vector of length 2. Helps identify which column the counts came from. Default c('x','y')}

\item{m_prob_func}{Function that takes a dataframe with columns token, token_type, n.x, n.y, n_comparisons, u_prob, and returns a vector of m_probs}
}
\description{
Generates a dataframe with the total counts of each tokens across both datasets as well as the m and u probs
}
\examples{

dat_ceo <- readr::read_csv('https://tinyurl.com/2p8etjr6')
dat_alb <- readr::read_csv('https://tinyurl.com/2p8ap4ad')
t_dat <- token_links(
  dat_x = dat_ceo,
  dat_y = dat_alb,
  args_x = list(col_nms = 'coname'),
  args_y = list(col_nms = 'companyName'),
  token_types = 'company_name',
  token_index = '',
  suffix = c('ceo', 'alb')
)
results <- generate_all_tokens(t_dat$x$token_counts, t_dat$y$token_counts, t_dat$total_comparisons)


}
