% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_random_column}
\alias{generate_random_column}
\title{Generates a single column dataframe}
\usage{
generate_random_column(
  n = 1000,
  k_min = 1,
  k_max = 4,
  k = round(runif(min = k_min, max = k_max, n = n), 0),
  token_type = "company_name",
  token_counts = generate_token_counts(a_token_type = token_type)
)
}
\arguments{
\item{n}{Number of rows returned in dataframe.}

\item{k_min}{Minimum number of words.}

\item{k_max}{Maximum number of words.}

\item{k}{a vector positive integers of length N}

\item{token_type}{string to indicate the type of token}

\item{token_counts}{counts tokens of that type}
}
\description{
Generates a single column dataframe
}
