% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_random_dat}
\alias{generate_random_dat}
\title{generate_random dataframe}
\usage{
generate_random_dat(n = 1000)
}
\arguments{
\item{n}{number of rows to return}
}
\description{
generate_random dataframe
}
