% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_random_dat_pairs}
\alias{generate_random_dat_pairs}
\title{Generate and save data and a pair that is edited randomly}
\usage{
generate_random_dat_pairs(
  n = 25000,
  fn = file.path(data_dir, "generated_dataset.csv")
)
}
\arguments{
\item{n}{number of rows to generate}

\item{fn}{file name to save the data to}
}
\description{
Generate and save data and a pair that is edited randomly
}
