% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_token_counts}
\alias{generate_token_counts}
\title{Generates counts of tokens for a given token type, returns a datafrgenerate_random_column(ame with two columns token and n}
\usage{
generate_token_counts(a_token_type = "company_name")
}
\arguments{
\item{a_token_type}{as string/ Defaults to company_name}
}
\description{
Generates counts of tokens for a given token type, returns a datafrgenerate_random_column(ame with two columns token and n
}
