% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{get_paired_row_names}
\alias{get_paired_row_names}
\title{Returns a dataframe with two columns indicating the rows of each Df that are to be paired}
\usage{
get_paired_row_names(
  t_dat,
  x_tokens = t_dat$x$tokens,
  y_tokens = t_dat$y$tokens,
  tokens_to_keep = NULL,
  token_join_by = c("token", "token_type"),
  suffix = paste0(".", c(t_dat$x$suffix, t_dat$y$suffix)),
  min_token_u_prob = 7.84e-05
)
}
\arguments{
\item{t_dat, }{a list that is a t_dat object}

\item{x_tokens}{defaults as t_dat$x$tokens}

\item{y_tokens}{defaults as t_dat$y$tokens}

\item{tokens_to_keep}{dataframe indicate which tokens matterin the analysis. typically this is a filtered subset of t_dat$tokens_all. Default is NULL. If value is NULL we filter t_dat$tokens_all by min_token_u_prob and u_prob != 0}

\item{token_join_by}{defaults to  c("token", "token_type")}

\item{suffix}{defaults to paste0(".",c(t_dat$x$suffix, t_dat$y$suffix))}

\item{min_token_u_prob}{minimum u_prob of token to use as join}
}
\description{
Returns a dataframe with two columns indicating the rows of each Df that are to be paired
}
