% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_duplicate.R
\name{group_df_from_graph}
\alias{group_df_from_graph}
\title{comverts a igraph to a dataframe, giving a group name to every subgraph}
\usage{
group_df_from_graph(g, group_prefix = "g_")
}
\arguments{
\item{g}{a graph}

\item{group_prefix}{prefix for group name}
}
\description{
comverts a igraph to a dataframe, giving a group name to every subgraph
}
