% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_duplicate.R
\name{group_duplicates}
\alias{group_duplicates}
\title{given a single dataframe, will de-duplicate that dataframe and return either a dataframe with grouping names or a TokenLink::t_dat object with the data.}
\usage{
group_duplicates(
  dat_x_y,
  col_nms_x_y = colnames(dplyr::select_if(dat_x_y, is.character)),
  args_x_y = list(),
  max_total_comparisons = TOKEN_MAX_COMPARE_DEFAULT * 10,
  min_token_u_prob = 0.001,
  min_posterior_all_evidence = 0.001,
  min_posterior_positive_evidence_only = 0.01,
  n_chunks = TokenLink:::N_CHUNKS,
  min_posterior_for_graph = 0.93,
  group_prefix = "g_",
  node_prefix = "N_",
  return_all = TRUE,
  ...
)
}
\arguments{
\item{dat_x_y}{a data frame to look for duplicates in.}

\item{...}{passed to token_link function}
}
\description{
given a single dataframe, will de-duplicate that dataframe and return either a dataframe with grouping names or a TokenLink::t_dat object with the data.
}
