% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{joined_results}
\alias{joined_results}
\title{Returns a joined dataframe}
\usage{
joined_results(
  t_dat,
  include_row_numbers = FALSE,
  pairs_to_get = "all_evidence",
  link_col_nms = c("posterior"),
  orig_data_to_include = "matched",
  ...
)
}
\arguments{
\item{t_dat}{a t_dat object}

\item{include_row_numbers}{Boolean Do we include the row names. Default False}

\item{pairs_to_get}{either 'all_evidence' or 'positive_evidence' or a dataframe. Default 'all_evidence'}

\item{link_col_nms}{vector of column names to includ from the joining dataframe. Default 'Posterior'}

\item{orig_data_to_include}{either 'matched' or 'all' and determines what is returned from the x and y dataframes. Default 0.3}
}
\description{
Returns a joined dataframe
}
\examples{

dat_ceo <- readr::read_csv('https://tinyurl.com/2p8etjr6')
dat_alb <- readr::read_csv('https://tinyurl.com/2p8ap4ad')
token_links(
dat_x = dat_ceo,
dat_y = dat_alb,
args_x = list(col_nms = 'coname'),
args_y = list(col_nms = 'companyName'),
token_types = 'company_name',
token_index = '',
suffix = c('ceo', 'alb')
) |> find_posterior_positive_evidance_only() |>
find_posterior_all_evidance() |>
joined_results()

}
