% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{map_s2c}
\alias{map_s2c}
\title{Change a string to a vector}
\usage{
map_s2c(str, ...)
}
\arguments{
\item{str}{a string}

\item{...}{passed to s2c in purrr::map}
}
\description{
Change a string to a vector
}
