% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{maybe_add}
\alias{maybe_add}
\title{adds value to lst with the key nm if nm is not already in lst}
\usage{
maybe_add(lst, nm, val)
}
\arguments{
\item{lst}{a list}

\item{nm}{string that is a key}

\item{val}{a value to add to the list}
}
\description{
adds value to lst with the key nm if nm is not already in lst
}
\examples{
list(a = 1, b = 2) |> maybe_add('c', 4) |> maybe_add('a', 99)


}
