% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{maybe_do}
\alias{maybe_do}
\title{will apply func to x if bool is TRUE. Saves us from an ugly function}
\usage{
maybe_do(x, bool, func, ...)
}
\arguments{
\item{x}{anything really , just the first argument of func}

\item{bool}{Boolean. applies function if bool is True}

\item{func}{function that takes x and ...}

\item{...}{passed to func}
}
\description{
will apply func to x if bool is TRUE. Saves us from an ugly function
}
\examples{
 maybe_do(c(2, 2,100), TRUE, function(x){x*2})
 maybe_do(c(2, 2,100), FALSE, function(x){x*2})

}
