% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{random_edit}
\alias{random_edit}
\title{edit the characters and words of a string}
\usage{
random_edit(str, ...)
}
\arguments{
\item{str}{A string}

\item{...}{passed to random_edit_char, and random_edit_word}
}
\description{
edit the characters and words of a string
}
