% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{random_edit_char}
\alias{random_edit_char}
\title{Randomly edit some string, by adding deleteing or swapping characters randomly}
\usage{
random_edit_char(
  str,
  ...,
  prob_char = 0.02,
  prob_del = 0.02,
  prob_insert = 0.02,
  sub_in = letters
)
}
\arguments{
\item{str}{A string}

\item{...}{ignored}

\item{prob_char}{probability that a character gets edited}

\item{prob_del}{probability that a character gets deleted}

\item{prob_insert}{probability that a character gets insert}

\item{sub_in}{vector of characters to use for replacing}
}
\description{
Randomly edit some string, by adding deleteing or swapping characters randomly
}
