% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{random_edit_number}
\alias{random_edit_number}
\title{Edit a number maybe? by moving it some small amount from it origional value.}
\usage{
random_edit_number(
  x,
  prob_small_move = 0.03,
  num_type = as.integer,
  frac_move = 0.1,
  ...
)
}
\arguments{
\item{x}{a number, or convertible to a number}

\item{prob_small_move}{Fraction probability tha number will change. Default as 0.3}

\item{num_type}{Converts to a number type default to as.integer}

\item{frac_move}{what percentage to move if it is moved}

\item{...}{ignored}
}
\description{
Edit a number maybe? by moving it some small amount from it origional value.
}
