% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{random_edit_word}
\alias{random_edit_word}
\title{Edit the words of the string maybe., by randomly inserting or deleting a word from the string}
\usage{
random_edit_word(str, ..., prob_del = 0.05, prob_insert = 0.05)
}
\arguments{
\item{str}{A string}

\item{...}{ignored}

\item{prob_del}{probability that a word gets edited}

\item{prob_insert}{probability that a word gets inserted}
}
\description{
Edit the words of the string maybe., by randomly inserting or deleting a word from the string
}
