% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{read_replacements_token_type}
\alias{read_replacements_token_type}
\title{Reads in a replacement token file}
\usage{
read_replacements_token_type(token_type = NULL, data_dir = "helper")
}
\arguments{
\item{token_type}{String. Type of token you are looking for, if NULL or replacement file not found returns NULL}

\item{data_dir}{String. the directory where we look for the token replacement file}
}
\description{
Reads in a replacement token file
}
\examples{
read_replacements_token_type('company_name')
read_replacements_token_type('poop')

}
