% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{read_replacements_token_type_get_fn}
\alias{read_replacements_token_type_get_fn}
\title{get the name of the token replacement file}
\usage{
read_replacements_token_type_get_fn(
  token_type,
  data_dir = REPLACEMENT_TOKEN_DATA_DIR,
  file_name_pattern = REPLACEMENT_TOKEN_FN_GLUE
)
}
\arguments{
\item{file_name_pattern}{}
}
\description{
get the name of the token replacement file
}
