% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{reclin_pair_blocking}
\alias{reclin_pair_blocking}
\title{creates a pair blocking based on columns passed into blocking_var, then then it generates more pairs based on token,
that are the same with columns col_nms_x and col_nms_y.}
\usage{
reclin_pair_blocking(
  x,
  y,
  blocking_var,
  token_types,
  large = FALSE,
  add_xy = TRUE,
  chunk_size = 1e+07,
  col_nms_x = colnames(select_if(x, is_character)),
  col_nms_y = colnames(select_if(y, is_character)),
  min_token_u_prob = TOKEN_MIN_UPROB_DEFAULT,
  ...
)
}
\arguments{
\item{x}{dataframe}

\item{y}{dataframe}

\item{blocking_var}{vector of column names to block on, unlike reclin these columns are joined with 'or'}

\item{token_types}{vector of token types}

\item{large}{Passed to reclin::pair_blocking. Default FALSE}

\item{add_xy}{Boolean Default True. This is passed to reclin::pair_blocking, inside map_dfr}

\item{chunk_size}{passed to reclin::pair_blocking. Default 1E+07}

\item{col_nms_x}{passed to token_links}

\item{col_nms_y}{passed to token_links}

\item{min_token_u_prob}{passed to get_paired_row_names}

\item{...}{passed to token_links}
}
\description{
creates a pair blocking based on columns passed into blocking_var, then then it generates more pairs based on token,
that are the same with columns col_nms_x and col_nms_y.
}
