% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{refine_posterior}
\alias{refine_posterior}
\title{After generating probabilities for a list of pairs this will refine the probabilities}
\usage{
refine_posterior(
  p,
  x_dat,
  y_dat,
  weights_nm = "x_weight",
  priori_delta = 0.01,
  args_x = list(col_nms = "company_name"),
  args_y = list(col_nms = "company_name"),
  token_types = "company_name",
  token_index = "",
  suffix = TOKEN_SUFFIX_DEFAULT,
  token_join_by = TOKEN_TOKEN_TYPE_VEC
)
}
\arguments{
\item{p}{should be a dataframe like object with atleast 3 columns 'x', 'y' indicating the row names of the x_dat and y_dat that are being compared}

\item{x_dat}{dataframe to check}

\item{y_dat}{dataframe to check}

\item{weights_nm}{Name of the column in pairs object that contains some kind of score or probability, which will be re scaled to a probability}

\item{priori_delta}{passed to scale_to_prob. Default = 0.01}

\item{args_x}{passed to token_links. Default list(col_nms = 'company_name')}

\item{args_y}{passed to token_links. Default list(col_nms = 'company_name')}

\item{token_types}{passed to token_links. Default 'company_name'}

\item{token_index}{passed to token_links. Default is a blank string.}

\item{suffix}{passed to token_links. Default is TOKEN_SUFFIX_DEFAULT}
}
\value{
returns a dataframe with x and y columns as well as priori and posterior from the original dataframe 'p' containing pairs with priori
}
\description{
After generating probabilities for a list of pairs this will refine the probabilities
}
