% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{s2c}
\alias{s2c}
\title{Get a single character out of a string}
\usage{
s2c(i, str, ...)
}
\arguments{
\item{i}{an integer}

\item{str}{a string}

\item{...}{passed to substr}
}
\description{
Get a single character out of a string
}
\examples{
\dontrun{
s2c(i = 1, "data Doo Doo")
}
}
