% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_helpers.R
\name{token_replacement_generator}
\alias{token_replacement_generator}
\title{returns a vector of replacement tokens}
\usage{
token_replacement_generator(
  .v,
  replace_space_with = "_",
  replace_punc_with = "_"
)
}
\arguments{
\item{.v}{vector of strings}

\item{replace_space_with}{replace punctuation with: Default '_'}

\item{replace_punc_with}{replace space with: Default '_'}
}
\description{
returns a vector of replacement tokens
}
\examples{
  token_replacement_generator(c('One Two Three', 'ehhh bee See'))
}
