% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{tokenize_ations_m_u_prob}
\alias{tokenize_ations_m_u_prob}
\title{Joins two objects together that come back from the tokenize_ations function, joins the token counts together, and calculates the m and u probs for each token.}
\usage{
tokenize_ations_m_u_prob(x, y, ..., suffix = TOKEN_SUFFIX_DEFAULT)
}
\arguments{
\item{x}{list returned from tokenize_ations}

\item{y}{list returned from tokenize_ations}

\item{...}{ignored}

\item{suffix}{String vector of length 2, identifies which original dataframe a column in the result comes from. Default TOKEN_SUFFIX_DEFAULT}
}
\description{
Joins two objects together that come back from the tokenize_ations function, joins the token counts together, and calculates the m and u probs for each token.
}
\examples{
dat_ceo <- readr::read_csv('https://tinyurl.com/2p8etjr6')
dat_alb <- readr::read_csv('https://tinyurl.com/2p8ap4ad')
tokenize_ations_m_u_prob(
  tokenize_ations(dat_ceo, col_nms = 'coname', token_types = 'TT'),
  tokenize_ations(dat_alb, col_nms = 'companyName', token_types = 'TT')
)


}
