% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_helpers.R
\name{write_token_replacement}
\alias{write_token_replacement}
\title{Given a vector of strings this will create or append a file of replacement tokens}
\usage{
write_token_replacement(
  .v,
  token_type,
  append_mode = "best",
  create_usable_or_testing_file = "testing",
  ...
)
}
\arguments{
\item{.v}{vector of strings}

\item{token_type}{a string}

\item{...}{passed to read_replacements_token_type_get_fn}
}
\description{
Given a vector of strings this will create or append a file of replacement tokens
}
