% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{calc_m_prob}
\alias{calc_m_prob}
\title{calculates what m_prob should be, takes in a dataframe and returns a vector of same length}
\usage{
calc_m_prob(
  dat_token_info,
  min_m_prob = 0.99,
  max_m_prob = 0.999,
  log_base = 10,
  ...
)
}
\arguments{
\item{dat_token_info}{a dataframe with information about the tokens}

\item{min_m_prob}{minimum value of m_prob returned}

\item{max_m_prob}{maximum value of m_prob returned}

\item{log_base}{Number. Base of the log. Default 10}

\item{...}{is ignored}
}
\description{
calculates what m_prob should be, takes in a dataframe and returns a vector of same length
}
