% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{find_posterior}
\alias{find_posterior}
\title{appends dataframes with posteriors and returns it}
\usage{
find_posterior(
  t_dat,
  min_posterior_all_evidance = 0.01,
  min_posterior_positive_evidance_only = 0.3,
  ...
)
}
\arguments{
\item{t_dat}{tdat object}

\item{min_posterior_all_evidance}{passed to find_posterior_all_evidance as min_posterior. Default 0.01}

\item{min_posterior_positive_evidance_only}{passed to find_posterior_all_evidance as min_posterior. Default 0.3}

\item{...}{passed wholsale to both find_posterior_positive_evidance_only, and find_posterior_all_evidance}
}
\description{
appends dataframes with posteriors and returns it
}
