% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{find_posterior}
\alias{find_posterior}
\title{Appends dataframes with posteriors and returns it}
\usage{
find_posterior(
  t_dat,
  min_posterior_all_evidence = MIN_POSTERIOR_ALL_EVIDENCE,
  min_posterior_positive_evidence_only = MIN_POSTERIOR_POSITIVE_EVIDENCE_ONLY,
  ...
)
}
\arguments{
\item{t_dat}{tdat object}

\item{min_posterior_all_evidence}{passed to find_posterior_all_evidence as min_posterior. Default MIN_POSTERIOR_ALL_EVIDENCE}

\item{min_posterior_positive_evidence_only}{passed to find_posterior_positive_evidence_only as min_posterior. Default MIN_POSTERIOR_POSITIVE_EVIDENCE_ONLY}

\item{...}{passed wholsale to both find_posterior_positive_evidence_only, and find_posterior_all_evidence}
}
\description{
Appends dataframes with posteriors and returns it
}
\examples{
t_dat <-
  token_links(
     dat_x = readr::read_csv('https://tinyurl.com/2p8etjr6'),
     dat_y = readr::read_csv('https://tinyurl.com/2p8ap4ad'),
     args_x = list(col_nms = 'coname'),
     args_y = list(col_nms = 'companyName'),
     token_types = 'company_name',
     token_index = '',
     suffix = c('ceo', 'alb')
  )

  t_dat <- t_dat |> find_posterior()


}
