% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{find_posterior_all_evidance}
\alias{find_posterior_all_evidance}
\title{t_dat should have been run through find_posterior_positive_evidance_only first to create the list of candidates}
\usage{
find_posterior_all_evidance(
  t_dat,
  min_posterior = 0.01,
  token_join_by = c("token", "token_type"),
  positive_evidance_only = NULL,
  return_all = TRUE,
  priori_func = calculate_priori,
  ...
)
}
\arguments{
\item{t_dat}{t_dat object}

\item{min_posterior}{filter results below this threshold Default = 0.01}

\item{positive_evidance_only, }{NULL or a dataframe with at least two columns indicating the row_names to be checked in eash of the datasets}

\item{return_all}{if TRUE it returns the whole object if FALSE it just returns the dataframe. Default TRUE}

\item{...}{Not used}

\item{token_join__by}{what to join the all_tokens by. Default = c("token", "token_type"),}
}
\description{
t_dat should have been run through find_posterior_positive_evidance_only first to create the list of candidates
}
\examples{
token_links(
 dat_x = readr::read_csv('https://tinyurl.com/2p8etjr6'),
 dat_y = readr::read_csv('https://tinyurl.com/2p8ap4ad' ),
 args_x = list(col_nms = 'coname'),
 args_y = list(col_nms = 'companyName'),
 token_types = 'company_name',
 token_index = '',
 suffix = c('ceo', 'alb')
) |> find_posterior_positive_evidance_only() |>
find_posterior_all_evidance(return_all = FALSE)


}
