% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenify.R
\name{find_posterior_positive_evidence_only}
\alias{find_posterior_positive_evidence_only}
\title{Creates a subset of pairs to check in more detail.}
\usage{
find_posterior_positive_evidence_only(
  t_dat,
  min_posterior = 0.2,
  token_join_by = c("token", "token_type"),
  tokens_to_keep = NULL,
  min_token_u_prob = 7.84e-05,
  return_all = TRUE,
  priori_func = calculate_priori,
  ...
)
}
\arguments{
\item{t_dat}{t_dat object}

\item{min_posterior}{filter posterior results above this value. Default 0.2.}

\item{token_join_by}{vector column names that joins the tokens. Default c("token", "token_type")}

\item{tokens_to_keep}{NULL or dataframe with a list of tokens, and m_prob and u_prob for each token, in the case where it is NULL we use tokenized$tokens_all filtered by min_token_u_prob}

\item{min_token_u_prob}{a float. This is ignored if tokens_to_keep is not NULL.  Default 0.0000784.}

\item{return_all}{if TRUE it returns the whole object if FALSE it just returns the dataframe. Default TRUE}

\item{priori_func}{A function that will calculate the priori. Default calculate_priori}

\item{...}{not used}
}
\description{
Creates a subset of pairs to check in more detail.
}
\examples{
token_links(
 dat_x = readr::read_csv('https://tinyurl.com/2p8etjr6'),
 dat_y = readr::read_csv('https://tinyurl.com/2p8ap4ad' ),
 args_x = list(col_nms = 'coname'),
 args_y = list(col_nms = 'companyName'),
 token_types = 'company_name',
 token_index = '',
 suffix = c('ceo', 'alb')
) |> find_posterior_positive_evidence_only(return_all = FALSE)

}
